backend/requirements.txt
# Web框架
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# 数据库和ORM
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Redis缓存
redis==5.0.1
aioredis==2.0.1

# 认证和安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# 向量数据库
qdrant-client==1.7.0
sentence-transformers==2.2.2

# 知识图谱
neo4j==5.15.0
py2neo==2021.2.4

# AI模型集成
openai==1.3.7
anthropic==0.7.8
tiktoken==0.5.2

# 国产AI模型SDK
zhipuai==2.0.1
dashscope==1.17.1
qianfan==0.3.9
erniebot==0.6.2
deepseek==1.0.0
opencc==1.1.6

# 文档处理
python-docx==1.1.0
PyPDF2==3.0.1
python-magic==0.4.27
openpyxl==3.1.2
xlrd==2.0.1

# 数据处理和分析
pandas==2.1.4
numpy==1.25.2
scikit-learn==1.3.2

# HTTP客户端
httpx==0.25.2
aiohttp==3.9.1

# 配置管理
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# 日志和监控
loguru==0.7.2
prometheus-client==0.19.0

# 任务队列
celery==5.3.4
flower==2.0.1

# 测试框架
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# 开发工具
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# 数据验证
email-validator==2.1.0
phonenumbers==8.13.25

# 时间处理
python-dateutil==2.8.2
pytz==2023.3

# 文件处理
Pillow==10.1.0
imageio==2.31.6

# 加密
cryptography==41.0.8

# 环境变量
environs==10.0.0

# API文档
fastapi-users==12.1.2
fastapi-pagination==0.12.13